+++ testutils/g505_samples.go
@@ -1,0 +1,24 @@
+package testutils
+
+import "github.com/securego/gosec/v2"
+
+var (
+	// SampleCodeG505 - Blocklisted import SHA1
+	SampleCodeG505 = []CodeSample{
+		{[]string{`
+package main
+
+import (
+	"crypto/sha1"
+	"fmt"
+	"os"
+)
+
+func main() {
+	for _, arg := range os.Args {
+		fmt.Printf("%x - %s\n", sha1.Sum([]byte(arg)), arg)
+	}
+}
+`}, 1, gosec.NewConfig()},
+	}
+)
